requirements.txt
django>=3.2
django-password-policies-iplweb @ git+https://github.com/frankhood/django-password-policies-iplweb.git@1.0.0#egg=django-password-policies-iplweb
djangorestframework